libnetdata/os: move egress-interface lookup out of stream-sender
The getsockname + getifaddrs interface match is OS-specific and does not belong
in the otherwise portable stream-sender.c. Move it behind a portable API,
os_socket_egress_interface(), in libnetdata/os/socket_egress_interface.{c,h}
(with the no-getifaddrs / Windows fallback), so stream-sender.c only stamps the
_net_default_iface label. The unit test moves with it as
os_socket_egress_interface_unittest(), run from the -W unittest path.
