+++ src/daemon/main.c
@@ -451,6 +451,7 @@ int netdata_main(int argc, char **argv) {
                             if (ctx_unittest()) return 1;
                             if (query_plan_unittest()) return 1;
                             if (uuid_unittest()) return 1;
+                            if (os_socket_egress_interface_unittest()) return 1;
                             if (dyncfg_unittest()) return 1;
                             if (eval_unittest()) return 1;
                             if (duration_unittest()) return 1;
