+++ src/libnetdata/os/os.h
@@ -24,6 +24,7 @@
 #include "uuid_generate.h"
 #include "setenv.h"
 #include "hostname.h"
+#include "socket_egress_interface.h"
 #include "os-freebsd-wrappers.h"
 #include "os-macos-wrappers.h"
 #include "os-windows-wrappers.h"
